refactor: MovieListItem Rule 분리
