+++ app/src/main/java/woowacourse/movie/domain/model/MovieListItemRule.kt
@@ -1,0 +1,5 @@
+package woowacourse.movie.domain.model
+
+interface MovieListItemRule<out T> {
+    fun whenIndex(index: Int): Class<out T>
+}
