+++ app/src/main/java/woowacourse/movie/view/movies/MoviesContract.kt
@@ -1,5 +1,7 @@
 package woowacourse.movie.view.movies
 
+import woowacourse.movie.domain.model.MovieListItem
+
 interface MoviesContract {
     interface View {
         fun showMovies(movies: List<MovieListItem>)
