+++ app/src/main/java/woowacourse/movie/view/movies/MoviesPresenter.kt
@@ -3,22 +3,35 @@ package woowacourse.movie.view.movies
 import woowacourse.movie.data.DummyAdvertisement
 import woowacourse.movie.data.DummyMovie
 import woowacourse.movie.domain.model.Movie
+import woowacourse.movie.domain.model.MovieListItem
+import woowacourse.movie.domain.model.MovieListItemRule
+import woowacourse.movie.domain.model.MovieListItemRuleImpl
 
 class MoviesPresenter(
     private val view: MoviesContract.View,
 ) : MoviesContract.Presenter {
+    private val movieListItemRule: MovieListItemRule<MovieListItem> = MovieListItemRuleImpl()
+
     override fun loadData() {
         view.showMovies(buildListWithAds(DummyMovie.dummyMovie))
     }
 
-    private fun buildListWithAds(movies: List<Movie>): List<MovieListItem> {
-        val result = mutableListOf<MovieListItem>()
-        movies.forEachIndexed { index, movie ->
-            result.add(MovieListItem.MovieItem(movie))
-            if ((index + 1) % 3 == 0) {
-                result.add(MovieListItem.AdItem(DummyAdvertisement.advertisement))
+    private fun buildListWithAds(movies: List<Movie>): List<MovieListItem> =
+        buildList {
+            var movieSize = 0
+            var index = 0
+            while (movieSize < movies.size) {
+                val clazz = movieListItemRule.whenIndex(index)
+                when (clazz) {
+                    MovieListItem.MovieItem::class.java -> {
+                        add(MovieListItem.MovieItem(movies[movieSize]))
+                        movieSize++
+                    }
+                    MovieListItem.AdItem::class.java -> {
+                        add(MovieListItem.AdItem(DummyAdvertisement.advertisement))
+                    }
+                }
+                index++
             }
         }
-        return result
-    }
 }
